fix worksheet intakeprofile relationship so delete button works now

+++ client-side/js/index.js
@@ -95,24 +95,27 @@ function renderWorksheetFromProfile(worksheets) {
 function deleteWorksheet() {
   console.log('getting to delete???');
   const deleteButton = document.querySelector('.delete__worksheet');
-  deleteButton.addEventListener("click", (event) => {
-    console.log("hello, anyone home?");
-    const worksheetId = event.target.parentElement.querySelector('.worksheetId').value;
-    console.log(worksheetId);
-    apiActions.deleteRequest('http://localhost:8080/profile/22/deleteWorksheet/' + worksheetId,
-     (profileInfo) => {
-       console.log(profileInfo),
-       apiActions.getRequest("http://localhost:8080/intake_profile", (user) => {
-        app.innerHTML = userWelcome(user);
-       });
-      //  navUserProfile();
-      // toProfilePageFromLogin();
-      // app.innerHTML = userWelcome(profileInfo);
-      // worksheets = profile[0].worksheets;
-      // renderWorksheetFromProfile(worksheets);
-      // deleteWorksheet();
+  if (deleteButton != null) {
+    deleteButton.addEventListener("click", (event) => {
+      console.log("hello, anyone home?");
+      const worksheetId = event.target.parentElement.querySelector('.worksheetId').value;
+      console.log(worksheetId);
+      apiActions.deleteRequest('http://localhost:8080/profile/22/deleteWorksheet/' + worksheetId,
+       (profileInfo) => {
+         console.log(profileInfo),
+         apiActions.getRequest("http://localhost:8080/intake_profile", (user) => {
+          app.innerHTML = userWelcome(user);
+         });
+        //  navUserProfile();
+        // toProfilePageFromLogin();
+        // app.innerHTML = userWelcome(profileInfo);
+        // worksheets = profile[0].worksheets;
+        // renderWorksheetFromProfile(worksheets);
+        // deleteWorksheet();
+      });
     });
-  });
+  }
+
 }
 
 function header() {
